# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Composer Web API",
    "version": "0.10.0",
    "description": "Auto-generated from src/web-server.ts routes. Components seeded from runtime schemas."
  },
  "servers": [
    {
      "url": "https://api.example.com"
    }
  ],
  "paths": {
...
    "/api/queue": {
      "get": {
        "summary": "Auto-generated from route definition",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "ComposerApiKey": []
          }
        ]
      },
      "post": {
        "summary": "Auto-generated from route definition",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "ComposerApiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ComposerApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-composer-api-key",
        "description": "Value of MAESTRO_WEB_API_KEY. Authorization: Bearer <key> is also accepted."
      }
    },
...
  }
}